Modify sys_write to print to console
sys_write now checks if the file descriptor argument is 1 (STDOUT),
in which case it prints to the console. Also we check if a pointer
referred to in a user program is valid by ensuring it is in the
user virtual address space and that the virtual address is mapped
to physical memory.