Fix line endings and add generic pintos-gdb
Windows CRLF endings were used for the last push, which spoils the
git diff functionality (since every line is different). The line
endings are once again LF (for Linux). Changed the line endings based
on https://stackoverflow.com/questions/1889559/git-diff-to-ignore-m

Also we added the generic pintos-gdb file so it is retained in utils