Fix bug with opening files and syscalls
Need to allow kernel to handle page faults as well and not only user
programs because a syscall might trigger a page fault if the page
has not been loaded yet.
Cannot simply use pagedir_get_page to check if a virtual address is
valid because it may not have been mapped yet. Need to check if there
is an entry for the virtual address in the supplemental page table.