Fix stack growth from syscalls
We added and modified the functions to check for a valid pointer
in syscall to now determine if the access indicates an attempt
to grow the stack, and if so, to add to the supplemental page
table.

The page fault handler is updated with this change and now handles
stack growth attempts. The code is very haphazard and dirty and
there is repeated code - these should be cleaned soon.