Implement memory mapped files
We added a hashmap to the thread that contains the list of mmap
map ids and the necessary information to manage these files.

Also, the mmap and munmap syscalls were properly implemented in
syscall, so the files are mapped to virtual memory if the inputs
do not fail (there is sufficient contiguous space). Files are
munmapped on process exit automatically, similar to the file
closing we implemented in last assignment.

The addition of mmap file pages to the supplemental page table
and loading of the pages upon page fault is also implemented,
similar to what has been done in load_exec.