Change thread in evict_page
Need to be checking the supplemental page table of the thread that is
currently holding the frame and not the thread that called evict page.
Also changed frame_free_page to memset the page to 0xcc to make it
easier to catch bugs where threads try to access pages after freeing
them.