Fix read/write bugs
Changed syscall to return the number of bytes read no matter what
(as opposed to returning 0 if the number of bytes read is less than
the size), since this size is used to allocate memory in tests.

Also, the page fault handler now calls the exit syscall as opposed
to killing the process on an attempt to write to a read-only file,
to handle this type of error message appropriately.