Add lock to prevent race conditions when loading a frame being evicted
Added lock to supp_page_table_entry struct that needs to be acquired
by load_swap if the page is in the process of getting evicted. This is
to prevent a race condition from occuring when a process tries to load
back a page that has not been completely evicted yet.