Implement file buffer cache with global lock
I added cache.h and cache.c, which create a static cache table,
and have functions to bring data to cache and to write it back
to disk. The eviction policy currently is CLOCK.

The whole process is currently in a global lock, so this will have
to be changed into per-entry (eviction and read-write) locks.