Remove global locks
I removed the global filesystem lock, and thus had to add several
lines to functions to ensure that the contents of the cache entry
is what it is expected to be (and if not, keep looping until it
is).

Also, I had to add an eviction lock to the cache entry that I
acquire and release on loading a file from disk and on eviction to
disk. I am not sure why I need this - and why I can't just writer
acquire the read/write lock.