Implement initial changes for subdirectories
A normal file vs directory flag and a parent directory sector
variable were added to the inode_disk struct. All the relevant
filesys functions were modified to take in a current directory,
and the calls to this directory were changed to pass that in
from the thread's current directory (a new variable added to
the thread struct).

Also the syscalls for inumber and isdir were implemented, and
these functions were updated in the corresponding files. No extra
test cases pass though - still at 35 failing.