Implement filesys syscalls for mkdir, chdir, readdir
Many changes were made across files to implement the syscalls.
While the directories can be treated as files for much of it,
there are some more nuanced changes that need to be made. A parser
was made to handle the full path names and convert them to a dir
struct.

Currently all the persistence tests and four of the filesys tests
fail.